requirements.txt
mysql-connector-python
Pillow
SpeechRecognition
gTTS
google-cloud-vision
pyttsx3
Django
Django_extensions
Ipython
pygame
tensorflow==1.15.0
opencv-python
numpy
matplotlib
django
django_extensions
keras